requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
yt-dlp==2023.11.16
openai==1.3.7
pytest==7.4.3
python-dotenv==1.0.0